Clarify: Chess Cuernavaca isn’t public domain in all places
However, Chess Cuernavaca is still open source in places where copyright
law allows the images of characters in a typeface to be copyrighted.

As per Eltra Corp. v. Ringer: 'typeface' can be defined as a set of
letters, numbers, or other symbolic characters, whose forms are related
by repeating design elements consistently applied in a notational system
and are intended to be embodied in articles, whose intrinsic utilitarian
function is for use in composing text or other cognizable combinations
of characters.